Moving from extractors to converters. Fixing loop variable scope

+++ src/main/java/org/jtwig/functions/impl/NumberFunctions.java
@@ -3,20 +3,21 @@ package org.jtwig.functions.impl;
 import org.jtwig.functions.annotations.JtwigFunction;
 import org.jtwig.functions.annotations.Parameter;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 
 public class NumberFunctions {
     @JtwigFunction("number_format")
-    public String numberFormat (@Parameter Object number, @Parameter Integer fractionDigits, @Parameter String decimalSeparator, @Parameter String groupingSeparator) {
+    public String numberFormat (@Parameter Object number, @Parameter BigDecimal fractionDigits, @Parameter String decimalSeparator, @Parameter String groupingSeparator) {
         number = (number == null) ? 0 : number;
 
         DecimalFormat numberFormat = new DecimalFormat();
         DecimalFormatSymbols decimalFormatSymbols = numberFormat.getDecimalFormatSymbols();
 
         if (fractionDigits != null) {
-            numberFormat.setMaximumFractionDigits(fractionDigits);
-            numberFormat.setMinimumFractionDigits(fractionDigits);
+            numberFormat.setMaximumFractionDigits(fractionDigits.intValue());
+            numberFormat.setMinimumFractionDigits(fractionDigits.intValue());
         }
 
         if (decimalSeparator != null && !decimalSeparator.isEmpty())
@@ -36,12 +37,12 @@ public class NumberFunctions {
     }
 
     @JtwigFunction("number_format")
-    public String numberFormat (@Parameter Object number, @Parameter Integer fractionDigits, @Parameter String decimalSeparator) {
+    public String numberFormat (@Parameter Object number, @Parameter BigDecimal fractionDigits, @Parameter String decimalSeparator) {
         return numberFormat(number, fractionDigits, decimalSeparator, null);
     }
 
     @JtwigFunction("number_format")
-    public String numberFormat (@Parameter Object number, @Parameter Integer fractionDigits) {
+    public String numberFormat (@Parameter Object number, @Parameter BigDecimal fractionDigits) {
         return numberFormat(number, fractionDigits, null, null);
     }
